Clippy fixes and add add features 'primitive' and 'instantiate' to ado_base
