+++ contracts/andromeda_wrapped_cw721/Cargo.toml
@@ -26,7 +26,7 @@ schemars = "0.8.3"
 cw-storage-plus = "0.9.1"
 cw721 = "0.9.1"
 andromeda-protocol = { version = "0.1.0", path = "../../packages/andromeda_protocol" }
-ado-base = { path = "../../packages/ado_base", version = "0.1.0" }
+ado-base = { path = "../../packages/ado_base", version = "0.1.0", features = ["instantiate"] }
 cw721-base = { version = "0.9.1", features = ["library"] }
 common = { version = "0.1.0", path = "../../packages/common" }
