+++ packages/ado_base/src/lib.rs
@@ -1,9 +1,12 @@
 mod execute;
+#[cfg(feature = "instantiate")]
 mod instantiate;
 #[cfg(test)]
 pub mod mock_querier;
 #[cfg(feature = "modules")]
 pub mod modules;
+#[cfg(feature = "primitive")]
+pub mod primitive;
 mod query;
 pub mod state;
 #[cfg(feature = "withdraw")]
