+++ packages/ado_base/src/state.rs
@@ -10,6 +10,7 @@ pub struct ADOContract<'a> {
     pub owner: Item<'a, Addr>,
     pub operators: Map<'a, &'a str, bool>,
     pub ado_type: Item<'a, String>,
+    #[cfg(feature = "primitive")]
     pub primitive_contract: Item<'a, Addr>,
     #[cfg(feature = "modules")]
     pub module_info: Map<'a, &'a str, Module>,
@@ -27,6 +28,7 @@ impl<'a> Default for ADOContract<'a> {
             owner: Item::new("owner"),
             operators: Map::new("operators"),
             ado_type: Item::new("ado_type"),
+            #[cfg(feature = "primitive")]
             primitive_contract: Item::new("primitive_contract"),
             #[cfg(feature = "modules")]
             module_info: Map::new("andr_modules"),
